Fixed 8 bugs
After edit parking : coming back to the details screen doesn’t show the changes
Clicking on the highlighted doesn’t change but clicking on the back button on the top of the screen not a problem.
Delete parking is making the app crash… the delete is not getting updated in the list. Will send the video for it
Create account is breaking and the details are not saved in the DB
Password update changes are not getting changed in the DB
After changing the password. I’m not able to login with either the old or new password
Using old password is also not working
Kill the app and then try logging in and the old password worked
Able to login to the same user after deleting it. Only after a sign out and then killing the app the deleted user is not recognized
Note : The user is getting deleted from the DB